+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ___Note:__ Yet to be released changes appear here._
 
 * `FIX`: show default context menu for text inputs in diagram ([#5803](https://github.com/camunda/camunda-modeler/issues/5803))
 * `FEAT`: show warning on selected version and connection version mismatch ([#5652](https://github.com/camunda/camunda-modeler/issues/5652))
+* `CHORE`: track variable outline panel open/close events ([#5878](https://github.com/camunda/camunda-modeler/pull/5878))
 * `DEPS`: update to `@camunda/linting@3.49.0`